requirements.txt
asgiref==3.6.0
cloudinary==1.33.0
Django==3.2.19
django-cloudinary-storage==0.3.0
django-filter==23.1
djangorestframework==3.14.0
Pillow==9.5.0
pytz==2023.3
sqlparse==0.4.4
